Repository work was corrected. Catching of edge cases was inserted to IpToDecimalMapper. Readme.md was created. Redundant lombok annotations were removed from GeoIp.class
