+++ src/main/java/com/example/geoip/model/GeoIp.java
@@ -11,34 +11,22 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 @Getter
 @Setter
-@AllArgsConstructor
-@NoArgsConstructor
 @EqualsAndHashCode
-@ToString
 @Entity
 @Table(name = "geo_ip")
 public class GeoIp {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(name = "start")
     private Long start;
-    @Column(name = "end")
     private Long end;
-    @Column(name = "country_code")
     private String countryCode;
-    @Column(name = "country_name")
     private String countryName;
-    @Column(name = "region_name")
     private String regionName;
-    @Column(name = "city_name")
     private String cityName;
-    @Column(name = "latitude")
     private String latitude;
-    @Column(name = "longitude")
     private String longitude;
 }
