+++ src/main/java/com/example/geoip/repository/GeoIpRepository.java
@@ -7,6 +7,9 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface GeoIpRepository extends JpaRepository<GeoIp, Long> {
-    @Query(value = "from GeoIp g where ?1 between g.start and g.end")
-    GeoIp findGeoIp(Long ipAsDecimal);
+//    @Query(value = "from GeoIp g where ?1 between g.start and g.end")
+//    GeoIp findGeoIp(Long ipAsDecimal);
+    GeoIp findGeoIpByStartIsLessThanEqualAndEndIsGreaterThanEqual(Long ipAsDecimal, Long ipRepeat);
+
+
 }
