+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/geoip
-spring.datasource.username=root
-spring.datasource.password=whatdoyoudoforaliving
+spring.datasource.url=YOUR_DB_URL
+spring.datasource.username=YOUR_DB_USERNAME
+spring.datasource.password=YOUR_DB_PASSWORD
 
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.format_sql=true
-#spring.liquibase.change-log=classpath:db/changelog/db.changelog-master.yaml
