+++ src/main/resources/init/start.sql
@@ -11,7 +11,7 @@ CREATE TABLE `geo_ip` (
                           PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=3080146 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 
-LOAD DATA LOCAL INFILE 'ABSOLUTE_PATH_TO_FILE_ip2location.CSV'
+LOAD DATA LOCAL INFILE 'ABSOLUTE_PATH_TO_FILE_ip2location.csv'
 INTO TABLE geo_ip
 FIELDS TERMINATED BY ','
 ENCLOSED BY '"'
